+++ onmt/scorers/__init__.py
@@ -1,5 +1,5 @@
 
-"""Module for dynamic data transfrom."""
+"""Module for dynamic scoring"""
 import os
 import importlib
 
@@ -20,7 +20,6 @@ def get_scorers_cls(metric_names):
 
 
 __all__ = ["get_scorers_cls", "build_scorers"]
-#            "load_transforms", "save_transforms", "TransformPipe"]
 
 
 def register_scorer(metric):
