+++ onmt/scorers/ter.py
@@ -12,8 +12,8 @@ class TerScorer(Scorer):
         super().__init__(opts)
 
     def compute_score(self, preds, texts_refs):
-        try:
+        if len(preds) > 0:
             score = corpus_ter(preds, [texts_refs]).score
-        except Exception:
+        else:
             score = 0
         return score
