+++ onmt/trainer.py
@@ -14,7 +14,7 @@ import traceback
 
 import onmt.utils
 from onmt.utils.logging import logger
-from onmt.translate.translator import ScoringPreparator
+from onmt.translate.utils import ScoringPreparator
 from onmt.scorers import get_scorers_cls, build_scorers
